Update: Reconstruct modules.py, and update package path.
